chore: rename DatasetContainer to Dataset
