+++ app/components/App.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import DatasetContainer from '../containers/DatasetContainer'
+import Dataset from '../components/Dataset'
 
 // App is the main component and currently the only view
 // Everything must flow through here
@@ -7,7 +7,7 @@ export default class App extends React.Component {
   render () {
     return (
       <div id='app'>
-        <DatasetContainer />
+        <Dataset />
       </div>
     )
   }
