--- app/scss/_dataset-container.scss
@@ -1,193 +1,0 @@
-////////////////////////////
-//    Dataset Container   //
-////////////////////////////
-
-#dataset-container {
-  display: flex;
-  flex-direction: column;
-  flex-grow: 1;
-  height: 100%;
-
-  .header {
-    height: 50px;
-    background-color: #315379;
-    color: #F8F2F2;
-    cursor: default;
-    flex-direction: row;
-    flex-grow: 0;
-    flex-shrink: 0;
-  }
-
-  .header-column {
-    display: inline-block;
-    height: 100%;
-    border-right: 1px solid #484d65;
-    padding: 10px;
-
-    &:hover {
-      background-color: #416588;
-    }
-
-    &.expanded {
-      border-right: none;
-    }
-
-    &.expanded:hover {
-      background-color: transparent;
-    }
-  }
-
-  .columns {
-    display: flex;
-    flex-direction: row;
-    flex: 1;
-
-    #sidebar {
-      position: relative;
-
-      .resize-handle {
-        position: absolute;
-        right: -5px;
-        top: 0px;
-        height: 100%;
-        width: 10px;
-        cursor: ew-resize;
-      }
-    }
-
-    .content {
-      background-color: #d8d8d8;
-      flex: 1;
-
-      .overlay {
-        position: absolute;
-        height: 100%;
-        width: 100%;
-        background-color: #000;
-        opacity: 0.4;
-      }
-    }
-  }
-
-  .current-dataset {
-    width: 250px;
-    position: relative;
-    z-index:10;
-    display:flex;
-
-    img {
-      margin-right:10px;
-    }
-
-    .label {
-      font-size: 0.75rem;
-      font-weight: 300;
-      line-height: 0.8rem;
-    }
-
-    .name {
-      font-weight: 600;
-      line-height: 1.5rem;
-    }
-
-    .arrow {
-      width: 0px;
-      height: 0px;
-      border-left: 5px solid transparent;
-      border-right: 5px solid transparent;
-      position: absolute;
-
-      &.expand {
-        border-top: 5px solid #fff;
-        right: 15px;
-        top: 24px;
-      }
-
-      &.collapse {
-        border-bottom: 5px solid #383838;
-        right: 15px;
-        top: 24px;
-      }
-    }
-
-    &.expanded {
-      color: #383838;
-    }
-  }
-
-  .dataset-list {
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    left: 0;
-    width: 250px;
-    background-color: #fff;
-    padding-top: 50px;
-  }
-}
-
-////////////////////////////
-//     Dataset Sidebar    //
-////////////////////////////
-
-#dataset-sidebar {
-  height: 100%;
-  width: 100%;
-
-  .sidebar-row {
-    width: 100%;
-    font-size: .9rem;
-    border-bottom: 1px solid #979797;
-    &:last-child {
-        border-right: 0;
-    }
-  }
-
-  #tabs {
-    width: 100%;
-    height: 30px;
-    background-color: #FFF;
-    cursor: default;
-
-    .tab {
-      width: 50%;
-      display: inline-block;
-      height: 100%;
-      padding: 6px;
-      text-align: center;
-      border-right: 1px solid #979797;
-
-      &:last-child {
-        border-right: 0;
-      }
-
-      &.active {
-        border-bottom: 3px solid #0666D6;
-      }
-
-      &:hover {
-        background-color: #efefef;
-      }
-    }
-  }
-
-  .changes {
-    height: 30px;
-    padding: 6px;
-    background-color: #F6F8FA;
-  }
-
-  .file-row {
-    height: 60px;
-    padding: 10px;
-
-    .label {
-      font-weight: 600;
-      font-size: 1.2rem;
-    }
-
-    &:hover {
-      background-color: #efefef;
-    }
-  }
-}
