+++ app/scss/_dataset.old.scss
@@ -1,0 +1,377 @@
+//
+// SCSS for Dataset components
+// 
+// Body
+// BodyReadOnly
+// Commits
+// Dataset
+// DatasetButtonGroup
+// DatasetHeader
+// DatasetRouter
+// DeleteDataset
+// Export
+// Meta
+// MetadataViewer
+// MetaSummary
+// Overview
+// RenameDataset
+// Render
+// StructureViewer
+// Transform
+// Viz
+
+////////////////////////////
+//          Body          //
+////////////////////////////
+.body-load-more {
+  margin: 10px;
+}
+
+.body-load-more > button {
+  margin: 0 0 0 10px;
+}
+
+////////////////////////////
+//       BodyReadOnly     //
+////////////////////////////
+.body-read-only-wrap {
+  margin: auto;
+  margin-top: 40px;
+  width: 600px;
+  text-align: center;
+}
+
+.body-read-only-margin-top {
+  margin-top: 20px;
+}
+
+////////////////////////////
+//         Commits        //
+////////////////////////////
+.commits-wrap {
+  margin-top: 40px;
+}
+
+.commits-note-wrap {
+  margin: 40px 20px;
+}
+
+////////////////////////////
+//         Dataset        //
+////////////////////////////
+.dataset-wrap {
+  display: flex;
+  flex-direction: column;
+}
+
+.dataset-center {
+  width: 300px;
+  height: 300px;
+  margin: 10em auto;
+  top: 0;
+  left: 0;
+  bottom: 0;
+  right: 0;
+  text-align: center;
+}
+
+////////////////////////////
+//   DatasetButtonGroup   //
+////////////////////////////
+.dataset-button-group-wrap {
+  display: flex;
+  flex-direction: column;
+  padding: 20px;
+  width: 300px;
+}
+
+.dataset-button-group-buttons {
+  display: flex;
+  justify-content: space-between;
+  flex-direction: row-reverse;
+  flex-wrap: wrap;
+}
+
+.dataset-button-group-button {
+  margin-bottom: 20px;
+}
+
+////////////////////////////
+//      DatasetHeader     //
+////////////////////////////
+.dataset-header-wrap {
+  padding: 20px;
+}
+
+.dataset-header-links {
+  margin-top: 30px;
+}
+
+////////////////////////////
+//      DatasetRouter     //
+////////////////////////////
+.dataset-router-contents {
+  display: flex;
+  justify-content: space-between;
+  height: 100%;
+}
+
+.dataset-router-info {
+  flex: 1 1 960px;
+  max-width: 900px;
+}
+
+.dataset-router-buttons {
+  flex: 1 1 300px;
+  display: flex;
+  justify-content: flex-end;
+}
+
+////////////////////////////
+//      DeleteDataset     //
+////////////////////////////
+.delete-dataset-wrap {
+  margin-top: 40px;
+  padding-left: 20px;
+  padding-right: 20px;
+}
+
+.delete-dataset-buttons {
+  display: flex;
+  justify-content: space-between;
+}
+
+.delete-dataset-cancel {
+  color: #fff;
+  margin-top: 5px;
+  margin-left: 10px;
+}
+
+.delete-entire-dataset {
+  padding-right: 20px;
+  position: absolute;
+  bottom: 10px;
+}
+
+.delete-entire-dataset-button {
+  margin-right: 20px;
+}
+
+
+////////////////////////////
+//      ExportDataset     //
+////////////////////////////
+.export-dataset-radio {
+  margin-top: 20px;
+  height: 50px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: space-between;
+}
+
+.export-dataset-buttons {
+  position: 0;
+  display: flex;
+  margin: 40px 0;
+  align-items: baseline;
+  justify-content: space-between;
+}
+
+////////////////////////////
+//     MetadataViewer     //
+////////////////////////////
+.metadata-viewer-wrap {
+  padding: 20px;
+}
+
+.metadata-viewer-title {
+  margin-top: 30px;
+}
+
+.metadata-viewer-table-wrap {
+  width: 100%;
+  margin-top: 10px;
+  border-radius: 3px;
+  border: 1px solid #eee;
+  overflow: hidden;
+}
+
+.metadata-viewer-table {
+  width: 100%;
+}
+
+.metadata-viewer-row {
+  border-bottom: 1px solid #eee;
+}
+
+.metadata-viewer-key {
+  padding: 8px 5px;
+}
+
+.metadata-viewer-link {
+  font-size: 17px;
+}
+
+.metadata-viewer-info {
+  color: $primary-light;
+}
+
+.metadata-viewer-data {
+  text-align: right;
+  color: $primary-light;
+  padding-right: 15px;
+}
+
+////////////////////////////
+//       MetaSummary      //
+////////////////////////////
+.meta-summary-wrap {
+  display: flex;
+  flex-direction: column;
+  margin: 20px; 
+}
+
+.meta-summary-content {
+  display: flex;
+  justify-content: space-between;
+}
+
+.meta-summary-description {
+  margin-right: 20px;
+  padding-right: 20px;
+  max-width: 560px;
+}
+
+.meta-summary-tags {
+  flex: 0 0;
+}
+
+.meta-summary-theme {
+  display: inline-block;
+  margin-right: 20px;
+}
+
+////////////////////////////
+//        Overview        //
+////////////////////////////
+.overview-wrap {
+  display: flex;
+  flex-direction: column;
+}
+
+.overview-fields {
+  font-size: 14px;
+  margin-top: 20px;
+}
+
+.overview-value {
+  margin-left: 5px;
+}
+
+.overview-key {
+  color: $primary-muted;
+}
+
+.overview-open {
+  display: inline-block;
+  background: $primary;
+  color: rgba(0,0,0,0.6);
+  border: 1px solid rgba(0,0,0,0.2);
+  border-radius: 20px;
+  padding: 4px 8px;
+  font-size: 12px;
+  margin-left: 20px;
+}
+
+.overview-title-wrap {
+  display: flex;
+  align-items: center;
+  margin-top: 25px;
+}
+
+.overview-name {
+  margin: 0px;
+}
+
+.overview-flex {
+  margin-top: 40px;
+  display: flex;
+  align-items: baseline;
+}
+
+.overview-nav {
+  margin-left: 5px;
+  font-size: 14px;
+}
+
+.overview-schema-wrap {
+  padding: 20px;
+}
+
+////////////////////////////
+//      RenameDataset     //
+////////////////////////////
+.rename-dataset-wrap {
+  padding: 20px;
+}
+
+.rename-dataset-form-wrap {
+  margin-top: 40px;
+  color: $white;
+}
+
+.rename-dataset-buttons {
+  display: flex;
+  margin: 40px 0;
+  align-items: baseline;
+  justify-content: space-between;
+}
+
+////////////////////////////
+//     StructureViewer    //
+////////////////////////////
+.structure-viewer-wrap {
+  margin-bottom: 40px;
+  padding: 20px;
+}
+
+.structure-viewer-title {
+  margin-top: 30px;
+}
+
+.structure-viewer-table-wrap {
+  width: 100%;
+  margin-top: 10px;
+  border-radius: 3px;
+  border: 1px solid #eee;
+  overflow: hidden;
+}
+
+.structure-viewer-table {
+  width: 100%;
+}
+
+.structure-viewer-row {
+  border-bottom: 1px solid #eee;
+}
+
+.structure-viewer-key {
+  padding: 8px 5px;
+}
+
+////////////////////////////
+//        Transform       //
+////////////////////////////
+.transform-wrap {
+  height: 100%;
+}
+
+////////////////////////////
+//           Viz          //
+////////////////////////////
+.viz-wrap {
+  height: 100%;
+}
+
+
