+++ app/scss/style.scss
@@ -23,8 +23,6 @@
 @import "compare";
 @import "stylesheet";
 
-@import "dataset-container";
-
 #app, #app-container {
   height: 100%;
 }
